requirements.txt
selenium==4.21.0
webdriver_manager==4.0.1
python-telegram-bot==21.1.1